requirements.txt
click==8.1.3
Flask==2.1.1
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
passlib==1.7.4
PyJWT==2.6.0
Werkzeug==2.2.3
zipp==3.15.0
pymongo==4.0
mongoengine==0.24.0
flask_mongoengine==0.9.5
Flask-JWT-Extended==4.2.3
python-dotenv
